fix: commit 24/07/2024

+++ admin/shop/src/components/management/category/_sections/view/index.tsx
@@ -4,7 +4,7 @@ import useGetDetailCategory from "../../hooks/use-get-detail-category";
 import { CategoryProduct } from "@/types/category";
 import DetailCategoryTableData from "./components/table";
 import { useNavigate } from "react-router-dom";
-import { useAppDispatch } from "@/hooks/hooks";
+import { useAppDispatch, useAppSelector } from "@/hooks/hooks";
 import { deleteCategoryAsync } from "@/redux/slices/categorySlice";
 import { toast } from "sonner";
 
@@ -14,24 +14,29 @@ type Props = {
 
 const CategoryDetailView = ({ id }: Props) => {
   const { category } = useGetDetailCategory({ id });
+  const { currentUser } = useAppSelector((state) => state.auth);
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
   const handleDeleteCategory = async () => {
-    try {
-      await dispatch(deleteCategoryAsync(id));
-      toast.success("Xóa danh mục sản phẩm thành công");
+    if (currentUser?.isAdmin) {
+      try {
+        await dispatch(deleteCategoryAsync(id));
+        toast.success("Xóa danh mục sản phẩm thành công");
 
-      navigate(`/dashboard/management/category`);
-    } catch (error) {
-      console.log(error);
-      toast.error("Xóa danh mục sản phẩm thất bại");
+        navigate(`/dashboard/management/category`);
+      } catch (error) {
+        console.log(error);
+        toast.error("Xóa danh mục sản phẩm thất bại");
+      }
+    } else {
+      toast.error("Bạn không có quyền xóa danh mục sản phẩm");
     }
   };
   return (
     <>
       <Header
         title="Thông tin loại sản phẩm"
-        text1="Xóa loại sản phẩm"
+        text1={"Xóa loại sản phẩm"}
         text2="Quản lý loại sản phẩm"
         onClick={handleDeleteCategory}
       />
