+++ admin/shop/src/components/products/_sections/view/detail.tsx
@@ -4,7 +4,7 @@ import useGetDetailProduct from "../../hooks/use-get-detail-product";
 import DetailProductView from "../components/detail-product-view";
 import { toast } from "sonner";
 import { useNavigate } from "react-router-dom";
-import { useAppDispatch } from "@/hooks/hooks";
+import { useAppDispatch, useAppSelector } from "@/hooks/hooks";
 import { deleteProductAsync } from "@/redux/slices/productSlice";
 
 type Props = {
@@ -14,17 +14,23 @@ type Props = {
 const ProductDetailView = ({ id }: Props) => {
   const { product } = useGetDetailProduct({ id });
 
+  const { currentUser } = useAppSelector((state) => state.auth);
+
   const navigate = useNavigate();
   const dispatch = useAppDispatch();
   const handleDeleteProduct = async () => {
-    try {
-      await dispatch(deleteProductAsync(id));
-      toast.success("Xóa sản phẩm thành công");
+    if (currentUser?.isAdmin) {
+      try {
+        await dispatch(deleteProductAsync(id));
+        toast.success("Xóa sản phẩm thành công");
 
-      navigate(`/dashboard/product`);
-    } catch (error) {
-      console.log(error);
-      toast.error("Xóa sản phẩm thất bại");
+        navigate(`/dashboard/product`);
+      } catch (error) {
+        console.log(error);
+        toast.error("Xóa sản phẩm thất bại");
+      }
+    } else {
+      toast.error("Bạn không có quyền xóa sản phẩm");
     }
   };
 
@@ -41,8 +47,8 @@ const ProductDetailView = ({ id }: Props) => {
         href2={`/dashboard/product/`}
         breadcumbItem="Sản phẩm"
         breadcumbPage="Thông tin chi tiết sản phẩm"
-        linkBtn={`/dashboard/product/${id}/edit`}
-        title="Chỉnh sửa sản phẩm"
+        linkBtn={currentUser?.isAdmin && `/dashboard/product/${id}/edit`}
+        title={currentUser?.isAdmin && "Chỉnh sửa sản phẩm"}
       />
 
       <DetailProductView product={product!} />
